Dielectric to separate module
